Wooo some logging!
It's not much, but we have some pretty basic logging (at least to a text
file) I think I need to refactor some of my functions to run off bits so
that I can handle a lot of this logging stuff in a single location
instead of multiple. That is going to be next on the list. Additionally,
the logger is available, but I'm not sure what I'll want to log, so I'll
need to work that out too. #16