Wooo! Even more logging!
This is a pretty sizable update (at least by my standards) and
represents a good couple hours of work.

Most notably, I've hit some major points on #14 , #16 , and #17 . Theres
a shell for robust exception handling (though not fully implemented) as
well as logging to a text file. I've started on the in-game message log,
but it still needs some work.

Working on this logging forced me to do some minor refactoring, that I
think helped the code quite a bit. A couple improvements to make (mostly
not involved with logging):
_ default colors for entities - currently, changing the message log
color with change everything
_ entity names - its really boring calling everything entity
_ scrolling the message log
_ only rendering the bottom of the list (will probably mean flipping the
list then rendering from the bottom)
_ continued work on possible layouts
_ half-width font form UI - will be crucial