half size font?
I was unhappy with how the in-game log was looking, so I started poking
around at using a half width font (i.e. 6x12 instead of 12x12). I
couldn't find one, so I took an 8x12 and resized it to fit :( it doesn't
look the greatest. Ive asked zoinks to work on building me  a new one.

Part of this change was to change the actual tiles of the terminal to
6x12 tiles. This means that the map needs to be rendered so that each
character takes up two tiles. The current system works (whereby it's
coded as if its on a square grid, then render at 2*x) but may not be the
best, needs some investigation.