Basic AI Component
I've created a very basic AI component that moves an AI in a random direction. This is currently very rough and makes a wall fly around like a bat, but it's functional enough to start. I think I want AI to funtion like this:
Player Takes turn
AI Takes Turn
AI plans next turn

This will allow us to put the actual AI code after a turn is taken, but before we render the screen.