Expanding the AI
This is a rough expansion of the ai, where it waits for the player to take a turn before it moves. I ran into issues and fixed some bugs relating to reseting movment and speed to 0 after a move (else an npc would move in a single direction forever), self blocking, and a few other things. I think the next thing I want to implement is a turn counter and an action stack. I want to be able to designate movement speeds for certain entities and have that go on the stack of turn order.